requirements.txt
streamlit>=1.30
pandas>=2.0
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pyyaml>=6.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
pyarrow>=14.0
pytest>=7.0
